Enhance command line rolling params and allow import of files (#474)
* Fix log statements to use configured log instead of default stdout

* Making rolling command arguments optional.

These values can mostly be inferred if not specified. But for TotalChunks
there is a default value that is set in the config file. Migrated rolling
options from the BroStaticCfg to their own RollingStaticCfg struct.

Additionally, non-rolling databases can be converted to rolling databases
by importing new data into them.

* Added ability to specify individual files to import

* Refactor GetFiles and removeOldFilesFromIndex to take database argument